--- packages/core/primitives/set/BUILD.bazel
@@ -1,30 +1,0 @@
-load("//tools:defaults.bzl", "ts_library", "tsec_test")
-
-package(default_visibility = [
-    "//packages/core/primitives/signals:__subpackages__",
-])
-
-ts_library(
-    name = "set",
-    srcs = glob(
-        [
-            "**/*.ts",
-        ],
-    ),
-)
-
-tsec_test(
-    name = "tsec_test",
-    target = "set",
-    tsconfig = "//packages:tsec_config",
-)
-
-filegroup(
-    name = "files_for_docgen",
-    srcs = glob([
-        "*.ts",
-        "src/**/*.ts",
-    ]),
-    visibility = ["//visibility:public"],
-)
-
