--- packages/core/primitives/set/index.ts
@@ -1,1 +1,0 @@
-export {LiteSet, createLiteSet, addToLiteSet, removeFromLiteSet} from './src/set';
