+++ packages/core/primitives/signals/BUILD.bazel
@@ -14,9 +14,6 @@ ts_library(
             "**/*.ts",
         ],
     ),
-    deps = [
-      "//packages/core/primitives/set"
-    ]
 )
 
 tsec_test(
