+++ packages/core/primitives/signals/src/set.ts
@@ -6,6 +6,9 @@ const SET_ID = Symbol('SET_ID');
  * slower to create and use than Arrays. To work with a single array, LiteSet
  * stashes information on hidden props on the array items themselves.
  *
+ * NOTE: This also assumes that the item objects isn't used as a dict (since
+ * the index would appear in things like Object.keys)
+ *
  * Use the functions below to manipulate the set. Read from the set with
  * standard indexing syntax.
  */
@@ -41,7 +44,7 @@ function indexOf<T extends Object>(set: LiteSet<T>, item: T): number | undefined
 /** Add an item to the LiteSet. No-op if the item is already in the set. */
 export function addToLiteSet<T extends Object>(set: LiteSet<T>, item: T): void {
   if (indexOf(set, item) === undefined) {
-    const index = set.push(item as T) - 1;
+    const index = set.push(item) - 1;
     (item as unknown as LiteSetItem)[`__idx_for_set_${set[SET_ID]}`] = index;
   }
 }
@@ -52,13 +55,13 @@ export function removeFromLiteSet<T extends Object>(set: LiteSet<T>, item: T): v
   if (index === undefined) return;
 
   // Cleanup the stored index on the item.
-  delete (set[index] as LiteSetItem)[`__idx_for_set_${set[SET_ID]}`];
+  delete (set[index] as unknown as LiteSetItem)[`__idx_for_set_${set[SET_ID]}`];
 
   if (set.length > 1) {
     // Swap the last item into the deleted position
     const lastIndex = set.length - 1;
     set[index] = set[lastIndex];
-    (set[index] as LiteSetItem)[`__idx_for_set_${set[SET_ID]}`] = index;
+    (set[index] as unknown as LiteSetItem)[`__idx_for_set_${set[SET_ID]}`] = index;
   }
 
   // Truncate the array
